// src/app/shared/models/employee.model.ts
export interface Employee {
  id?: number;
  firstName: string;
  lastName: string;
  email: string;
  age: number;
  gender: string;
  maritalStatus: string;
  distanceFromHome: number;
  matricule: string;
  department: string;
  jobRole: string;
  jobLevel: number;
  businessTravel: string;
  educationField: string;
  education: number;
  monthlyIncome: number;
  dailyRate: number;
  hourlyRate: number;
  monthlyRate: number;
  percentSalaryHike: number;
  stockOptionLevel: number;
  jobSatisfaction: number;
  environmentSatisfaction: number;
  relationshipSatisfaction: number;
  jobInvolvement: number;
  workLifeBalance: number;
  performanceRating: number;
  overTime: boolean;
  numCompaniesWorked: number;
  totalWorkingYears: number;
  yearsAtCompany: number;
  yearsInCurrentRole: number;
  yearsSinceLastPromotion: number;
  yearsWithCurrManager: number;
  trainingTimesLastYear: number;
  active?: boolean;
  derniereProbabilite?: number;
  dernierNiveauRisque?: string;
}